+++ src/gql/helpers.js
@@ -1,5 +1,6 @@
 /* @flow */
 import {mapKeys, snakeCase, camelCase, some, includes} from 'lodash';
+import crypto from 'crypto';
 import Boom from 'boom';
 import {GraphQLInputObjectType} from 'graphql';
 
@@ -95,4 +96,35 @@ export function assertEnvironment(environments: Array<string>) {
  */
 export function hasRole(user:AuthUserType, role: string): boolean {
   return some(user.roles, ['name', role]);
+}
+
+export function decrypt(ciphertext: string): string {
+  if (! (process.env.CRYPTO_KEY && process.env.CRYPTO_ALGO)) {
+    throw Boom.badImplementation('Missing require environment variables to encrypt data');
+  }
+
+  let decrypted: string = '';
+  try {
+    const key: Buffer = new Buffer(process.env.CRYPTO_KEY, 'hex');
+    const decipher = crypto.createDecipher(process.env.CRYPTO_ALGO, key);
+    // $FlowBug crypto$Decipher.update is broken in Flow
+    decrypted = decipher.update(ciphertext, 'hex', 'utf8');
+    // $FlowBug crypto$Decipher.final is broken in Flow
+    decrypted += decipher.final('utf8');
+  } catch (err) {
+    decrypted = '';
+  }
+  return decrypted;
+}
+
+export function encrypt(plaintext: string): string {
+  if (! (process.env.CRYPTO_KEY && process.env.CRYPTO_ALGO)) {
+    throw Boom.badImplementation('Missing require environment variables to encrypt data');
+  }
+
+  const key = new Buffer(process.env.CRYPTO_KEY, 'hex');
+  const cipher = crypto.createCipher(process.env.CRYPTO_ALGO, key);
+  let ciphertext = cipher.update(plaintext, 'utf8', 'hex');
+  ciphertext += cipher.final('hex');
+  return ciphertext;
 }