+++ src/lib/heretic.js
@@ -3,24 +3,10 @@
 import Heretic from 'heretic';
 import {knex} from './knex';
 
-const heretic = new Heretic(process.env.AMQP_URL, knex, {
+export const heretic = new Heretic(process.env.AMQP_URL, knex, {
   socketOptions: {
     clientProperties: {
       Application: 'Workers',
     },
   },
 });
-
-export const worker = {
-  enqueue(queueName: string, payload: any) {
-    return heretic.enqueue(queueName, payload);
-  },
-
-  dequeue(queueName: string, concurrency: number, handler: Function) {
-    return heretic.dequeue(queueName, concurrency, handler);
-  },
-
-  retry(jobId: number) {
-    return heretic.retry(jobId);
-  },
-};