+++ src/web/pre/require-jwt-auth.js
@@ -1,3 +1,0 @@
-/* @flow */
-export const requireJWTAuth =
-  process.env.DISABLE_JWT_AUTH === 'true' ? false : 'jwt';